sw_services: xilsecure: Fix for MISRA C Violation Rule 10.3
MISRA C Rule 10.3 states: The value of an expression shall not be assigned to an object with a narrower essential type or of a different essential type category.

This patch fixes MISRA C 10.3 violation by changing the return type of XSecure_Init() to int

Acked-by: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_services/xilsecure/src/versal/xsecure_init.c
@@ -62,11 +62,11 @@ static u32 XSecure_RegisterTampIntHandler(void);
  *		- Returns error code on failure
  *
  *****************************************************************************/
-u32 XSecure_Init(void)
+int XSecure_Init(void)
 {
-	u32 Status = (u32)XST_FAILURE;
+	int Status = XST_FAILURE;
 
-	Status = XSecure_RegisterTampIntHandler();
+	Status = (int)XSecure_RegisterTampIntHandler();
 
 	return Status;
 }
